--- DTO/AuthResponse.java
@@ -1,10 +1,0 @@
-package org.example.turboaz.DTO;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AuthResponse {
-    private String message;
-}
