--- DTO/ConfirmOtpRequest.java
@@ -1,11 +1,0 @@
-package org.example.turboaz.DTO;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ConfirmOtpRequest {
-    private String email;
-    private String otp;
-}
