--- DTO/RegisterRequest.java
@@ -1,13 +1,0 @@
-package org.example.turboaz.DTO;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RegisterRequest {
-    private String name;
-    private String email;
-    private String password;
-    private String phone;
-}
