--- Entity/CarAd.java
@@ -1,34 +1,0 @@
-package org.example.turboaz.Entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "Car")
-@Getter
-@Setter
-@RequiredArgsConstructor
-public class CarAd {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-    private String description;
-    private String brand;
-    private String model;
-    private int year;
-    private int mileage;
-    private double price;
-    private String imageUrl;
-    private LocalDateTime createdAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-}
