--- Repository/UserRepository.java
@@ -1,12 +1,0 @@
-package org.example.turboaz.Repository;
-
-import org.example.turboaz.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
