--- Security/SecurityConfig.java
@@ -1,59 +1,0 @@
-package org.example.turboaz.Security;
-
-import lombok.RequiredArgsConstructor;
-import org.example.turboaz.Security.CustomUserDetailsService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final CustomUserDetailsService userDetailsService;
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.csrf(csrf -> csrf.disable())
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/api/auth/register").permitAll()
-                        .requestMatchers("/api/auth/confirm").permitAll()
-                        .requestMatchers("/api/cars/getall").permitAll()
-                        .requestMatchers("/api/cars/create").hasAnyRole("ADMIN" , "USER")
-                        .requestMatchers("/api/cars/delete").hasAnyRole("ADMIN" , "USER")
-                        .requestMatchers("/api/cars/update").hasAnyRole("ADMIN" , "USER")
-                        .requestMatchers("/api/cars/uploadImage/{carAdId}").hasAnyRole("ADMIN" , "USER")
-                        .requestMatchers("/uploads/**").permitAll()
-                        .requestMatchers("/api/cars/filter").permitAll()
-                        .requestMatchers("/api/admin/delete/car/{carAdId}").hasRole("ADMIN")
-                        .anyRequest().authenticated()
-                )
-                .httpBasic(Customizer.withDefaults());
-        return http.build();
-    }
-
-
-    @Bean
-    public AuthenticationManager authenticationManager(HttpSecurity http) throws Exception {
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setUserDetailsService(userDetailsService);
-        provider.setPasswordEncoder(passwordEncoder());
-
-        return http.getSharedObject(org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder.class)
-                .authenticationProvider(provider)
-                .build();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
