+++ src/main/java/org/example/turboaz/dto/AuthResponse.java
@@ -1,0 +1,10 @@
+package org.example.turboaz.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class AuthResponse {
+    private String message;
+}
