+++ src/main/java/org/example/turboaz/dto/ConfirmOtpRequest.java
@@ -1,0 +1,11 @@
+package org.example.turboaz.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ConfirmOtpRequest {
+    private String email;
+    private String otp;
+}
