+++ src/main/java/org/example/turboaz/dto/RegisterRequest.java
@@ -1,0 +1,13 @@
+package org.example.turboaz.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class RegisterRequest {
+    private String name;
+    private String email;
+    private String password;
+    private String phone;
+}
