+++ src/main/java/org/example/turboaz/dto/UpdateCarAdRequest.java
@@ -1,0 +1,15 @@
+package org.example.turboaz.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class UpdateCarAdRequest {
+    private String title;
+    private String description;
+    private double price;
+}
+
