+++ src/main/java/org/example/turboaz/entity/User.java
@@ -1,0 +1,36 @@
+package org.example.turboaz.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "User")
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "Name")
+    private String name;
+
+    @Column(unique = true , name = "Email")
+    private String email;
+
+    @Column(name = "Password")
+    private String password;
+
+    @Column(name = "Phone")
+    private String phone;
+
+    @Column(name = "Role")
+    private String role;
+
+    @Column(name = "isEnable")
+    private boolean isEnable;
+}
