+++ src/main/java/org/example/turboaz/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package org.example.turboaz.repository;
+
+import org.example.turboaz.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+}
